progress towards new thumbnail design
