+++ app/src/App.css
@@ -28,7 +28,7 @@
     --primary-font-color: #fafafa;
     --secondary-font-color: #333;
 
-    --border-color: #f5f5f53b;
+    --border-color: #f5f5f524;
 
     --analytics-gap: 7px;
 }
