+++ app/src/components/Sidebar/Sidebar.css
@@ -12,6 +12,8 @@
     padding: 0 10px 10px 10px;
     border-radius: 4px;
     border: 1px solid var(--border-color);
+    background-color: #3d485300;
+    background-image: linear-gradient(360deg, #252b3294, #0b0e12d4 74%);
     color: var(--primary-font-color);
     transition: width 1s linear;
